# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/weight": {
      "post": {
        "operationId": "WeightController_logWeight",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogWeightDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "weight"
        ]
      }
    },
    "/weight/history": {
      "get": {
        "operationId": "WeightController_getHistory",
        "parameters": [
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "weight"
        ]
      }
    },
...
    "/coach/clients": {
      "get": {
        "operationId": "CoachController_getClients",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coach"
        ]
      }
    },
...
    "/users/me/preferences": {
      "get": {
        "description": "Returns current preferences with defaults filled in.",
        "operationId": "UsersController_getPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferences object."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get personalization preferences",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "description": "Merges a partial UserPreferencesDto into the stored prefs.",
        "operationId": "UsersController_patchPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Updated preferences object."
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Patch personalization preferences",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/badges": {
      "get": {
        "description": "Earned badges include awardedAt; locked badges have awardedAt = null.",
        "operationId": "UsersController_getBadges",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Badge list."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List badges (earned + locked) for the caller",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/founding-number": {
      "get": {
        "description": "Returns 1-indexed join rank, total user count, and whether the user is a Founding Member (rank ≤ 1000).",
        "operationId": "UsersController_getFoundingNumber",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Founding-member info."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Founding-member rank",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/circle-stats": {
      "get": {
        "description": "How many users in the caller’s coach→student group trained today, plus group size.",
        "operationId": "UsersController_getCircleStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Circle stats."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stats for the caller’s circle",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/data-export": {
      "post": {
        "description": "Synchronously assembles the snapshot, persists a DataExportRequest row, and returns a handle. Fetch the payload at GET /users/me/data-export/:id.",
        "operationId": "UsersController_requestDataExport",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Export request handle."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request a GDPR/CCPA personal-data export",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/data-export/{id}": {
      "get": {
        "operationId": "UsersController_getDataExport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data export payload."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "description": "Export not found or not owned by caller."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Fetch a previously requested data export by id",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/account": {
      "delete": {
        "description": "Idempotent — a second call within the grace window returns the same scheduled_at. Use POST /users/me/account/cancel-deletion to undo.",
        "operationId": "UsersController_deleteAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deletion scheduled (or already scheduled)."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedule account deletion (30-day grace period)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/account/cancel-deletion": {
      "post": {
        "operationId": "UsersController_cancelDeletion",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deletion canceled."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          },
          "404": {
            "description": "No deletion scheduled."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a scheduled account deletion",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/account/deletion-status": {
      "get": {
        "operationId": "UsersController_deletionStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deletion status."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get deletion status (canonical path)",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/account/status": {
      "get": {
        "description": "Same body as GET /users/me/account/deletion-status.",
        "operationId": "UsersController_accountStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Deletion status."
          },
          "401": {
            "description": "Missing or invalid bearer token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get deletion status (mobile-friendly alias)",
        "tags": [
          "users"
        ]
      }
    },
...
    "/v1/coach/me/threads/{clientId}": {
      "get": {
        "operationId": "V1CoachController_getThread",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coach-v1"
        ]
      }
    },
    "/v1/coach/me/threads/{clientId}/messages": {
      "post": {
        "operationId": "V1CoachController_sendMessage",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "coach-v1"
        ]
      }
    },
    "/v1/coach/me/threads/{clientId}/draft": {
      "get": {
        "operationId": "V1CoachController_getDraft",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "coach-v1"
        ]
      },
      "post": {
        "operationId": "V1CoachController_saveDraft",
        "parameters": [
          {
            "name": "clientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/V1SaveDraftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "coach-v1"
        ]
      }
    },
...
  },
  "info": {
    "title": "Growth Project API",
    "description": "HTTP API for The Growth Project — coach/client SaaS platform covering auth, billing, messaging, check-ins, meal plans, workouts, habits, and admin/federation surfaces. Consumed by the React Native mobile app, the coach console BFF, and partner integrations.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "https://api.trygrowthproject.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Supabase-issued JWT. Send as `Authorization: Bearer <token>`. Tokens are verified against the Supabase JWKS in JwtAuthGuard."
      }
    },
    "schemas": {
...
      "LogWeightDto": {
        "type": "object",
        "properties": {}
      },
...
      "V1SendMessageDto": {
        "type": "object",
        "properties": {}
      },
      "V1SaveDraftDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}